[#13][#22] Rename webform component to form container
