+++ app/src/form/FormContainer.js
@@ -5,27 +5,20 @@ import { Formik } from 'formik';
 import { styles } from './styles';
 import { FormNavigation } from './support';
 import QuestionGroup from './components/QuestionGroup';
-import { transformForm, generateValidationSchema } from './lib';
+import { transformForm } from './lib';
 
 // TODO:: Allow other not supported yet
 // TODO:: Repeat group not supported yet
 // TODO:: Cascade not supported yet
 // TODO:: Geo not supported yet
-// TODO:: Validation for dependency question not supported yet
 
-const Webform = ({ forms, initialValues = {} }) => {
-  //TODO:: Rename Webform component
-  // TRY TO USE FIELD LEVEL VALIDATION
+const FormContainer = ({ forms, initialValues = {} }) => {
   const formRef = React.useRef();
   const [activeGroup, setActiveGroup] = React.useState(0);
 
   const formDefinition = React.useMemo(() => {
     return transformForm(forms);
   }, [forms]);
-
-  const validationSchema = React.useMemo(() => {
-    return generateValidationSchema(formDefinition);
-  }, [formDefinition]);
 
   return (
     <>
@@ -34,9 +27,6 @@ const Webform = ({ forms, initialValues = {} }) => {
           <Formik
             innerRef={formRef}
             initialValues={initialValues}
-            // validationSchema={validationSchema}
-            // validateOnChange={false}
-            // validateOnBlur={true}
             onSubmit={(values) => console.log(values)}
           >
             {({ setFieldValue, values }) => (
@@ -76,4 +66,4 @@ const Webform = ({ forms, initialValues = {} }) => {
   );
 };
 
-export default Webform;
+export default FormContainer;
