+++ app/src/form/index.js
@@ -1,1 +1,1 @@
-export { default as Webform } from './Webform';
+export { default as FormContainer } from './FormContainer';
