+++ app/src/pages/FormPage.js
@@ -1,18 +1,18 @@
 import React from 'react';
-import { Webform } from '../form';
+import { FormContainer } from '../form';
 import { BaseLayout } from '../components';
 import * as formDefinition from '../form/example-form.json';
 
-const WebformPage = ({ navigation, route }) => {
+const FormPage = ({ navigation, route }) => {
   const goBack = () => {
     navigation.navigate('FormAction', { ...route?.params });
   };
 
   return (
     <BaseLayout title={route?.params?.name} back={goBack}>
-      <Webform forms={formDefinition} initialValues={{}} />
+      <FormContainer forms={formDefinition} initialValues={{}} />
     </BaseLayout>
   );
 };
 
-export default WebformPage;
+export default FormPage;
